refactor(fcp): split FCPServer helpers
Extract listener, config registration, plugin connection, and persistent request operations into focused helpers to reduce FCPServer coupling while keeping its public API stable.
